Phase K Wave 9: 0x0044 REQUEST_TIME round-trip + 2 server fixes
Adds SectorRequestTimeTests covering the simplest in-sector
request/reply pair (HandleRequestTime is unconditional, no
player-state guards). Client sends a unique-per-run tick sentinel;
server echoes it back in CLIENT_SET_TIME along with server tick
fields. Coverage ratchet 17 -> 19.

Server fix 1 -- PlayerConnection.cpp HandleRequestTime sizeof(long)
over-read. Wire payload is 4 bytes on Win32 (sizeof(long)==4) but
the unmigrated Linux server read 8 bytes via *((long *) data),
pulling 4B of payload plus 4B of recv-buffer leak and echoing the
garbage back to the client. Cast to int32_t* per the Wave 7 pattern.

Server fix 2 -- AuxBase::m_Max_Buffer uninit (surfaced by the test
running 3rd in the sector suite). Only 4 of 57 AuxBase subclasses
explicitly set m_Max_Buffer; the other 53 inherit whatever heap
garbage was there. On Win32 the garbage happened to be large and
AddData's overflow guard passed through; on Linux x86_64 the
garbage is small, so every AddData hits the
"Error: Bufferoverflow in Aux!" printf branch. With 3+ players in
a sector the per-player serialisation flood saturates stdout and
blocks the sector tick from servicing UDP -- REQUEST_TIME never
reaches HandleRequestTime and the test times out at 90s.

Initialise m_Max_Buffer to ULONG_MAX in the base constructor.
Subclasses that know a tighter cap keep their explicit assignments
(AuxHulkIndex=1000, AuxShipIndex=10000, AuxMobIndex=2000,
AuxHarvestable=2000) and override the default. Per CLAUDE.md
server-integrity rules this is preservation-aligned: restores 15+
years of upstream behaviour (Win32 happened to have large garbage
there) without weakening any real bound -- caller-side buffer
sizing is the load-bearing safety mechanism, not this check.

41/41 integration tests green in 3m17s from clean compose down -v.

+++ server/src/PlayerConnection.cpp
@@ -1620,8 +1620,12 @@ void Player::HandleRequestTime(unsigned char *data)
 {
 	LogDebug("Received RequestTime packet\n");
 
-	// Set the client time
-	SendClientSetTime(*((long *) data));
+	// Phase K Wave 9: wire field is 4 bytes (Win32 sizeof(long)==4); reading
+	// `*((long *) data)` on Linux pulls 8 bytes — 4 of payload + 4 of whatever
+	// follows in the recv buffer — which then echoes garbage back to the client
+	// as its own "ClientSent" tick. Same class of fix as the Wave 7 sweep across
+	// PlayerConnection/PlayerClass/PlayerManager.
+	SendClientSetTime(*((int32_t *) data));
 }
